+++ PizzaStore.Domain/Models/Size.cs
@@ -3,11 +3,11 @@ namespace PizzaStore.Domain.Models
     public class Size
     {
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         public Size() { }
 
-        public Size(string name, double price)
+        public Size(string name, decimal price)
         {
             Name = name;
             Price = price;
